+++ resources/views/emails/banner-classified.blade.php
@@ -1,0 +1,9 @@
+@component('mail::message')
+Hello,
+
+Your advertisement was classified.
+
+Thanks,
+
+{{ config('app.name') }} Team
+@endcomponent